requirements.txt
tensorflow
keras-applications==1.0.8
keras-preprocessing==1.1.2
tenseal
pandas
numpy
prometheus_api_client
boto3
fastapi
uvicorn